Split to firmwareActions.js and rssiActions.js
